完成client向server发送单键值对的http请求功能
